fix: Fix MegaflixExtractor
