+++ src/pt/megaflix/build.gradle
@@ -15,6 +15,7 @@ ext {
 dependencies {
     implementation(project(':lib-mixdrop-extractor'))
     implementation(project(":lib-streamtape-extractor"))
+    implementation(project(":lib-playlist-utils"))
     // for mixdrop and megaflix
     implementation("dev.datlag.jsunpacker:jsunpacker:1.0.1")
 }
